rag-service/requirements.txt
accelerate==0.26.1
openai==0.28
elasticsearch==8.12.0
Flask==3.0.2
numpy==1.26.3
PyYAML==6.0.1
tokenizers==0.15.1
torch==2.2.0
tqdm==4.66.1
transformers==4.37.2
supabase==2.3.4